requirements.txt
flask
transformers
torch
sentencepiece
nltk
python-dotenv